deploy/klai-connector/pyproject.toml
[project]
name = "klai-connector"
version = "0.1.0"
description = "External source connectors for Klai Knowledge via Unstructured.io"
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.7.0",
    "unstructured>=0.16.0",
    "httpx>=0.28.0",
    "apscheduler>=3.10.0",
    "cryptography>=44.0.0",
    "gidgethub>=5.3.0",
    "PyJWT[crypto]>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["B008"]  # Depends() in FastAPI endpoint defaults is standard pattern

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
